Pulling up code in an abstract class
