+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/migration/SyntheticTokenAllowanceOwnerMigration.java
@@ -152,6 +152,8 @@ public class SyntheticTokenAllowanceOwnerMigration extends RepeatableMigration i
         if (streamFile == null) {
             return;
         }
+
+        // The services version 0.38.0 has the fixes this migration solves.
         if (streamFile.getHapiVersion().isGreaterThanOrEqualTo(HAPI_VERSION_0_37_0)
                 && executed.compareAndSet(false, true)) {
             var latestFile = recordFileRepository.findLatestBefore(streamFile.getConsensusStart());
